fix output attention
